back 0214am commit - token ~ update & delete compare

+++ Back_End/src/main/java/com/ssafy/wine/controller/FeedController.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -37,21 +39,23 @@ public class FeedController {
 	@Autowired
 	private FeedService feedService;
 
-	@Autowired
-	private UserController userController;
-
 	@ApiImplicitParams({
 			@ApiImplicitParam(name = "TOKEN", value = "로그인 성공 후 access_token", required = true, dataType = "String", paramType = "header") })
 	@ApiOperation(value = "Feed 추가 - rating 값으로 review 판단")
 	@PostMapping("/create")
 	public ResponseEntity<Object> createReview(@RequestBody FeedInputDto feedInput) {
 		try {
-			Long uid = userController.findUserById().getData().getUid();
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			Long uid = Long.parseLong(authentication.getName());
+			
 			Feed feed = feedService.create(uid, feedInput);
-			StringBuilder sb = new StringBuilder();
-			sb.append("User: ").append(feed.getUser().getEmail()).append("\n").append("Feed_Id: ").append(feed.getFid())
-					.append("\nFeed 작성완료");
-			return new ResponseEntity<Object>(sb, HttpStatus.OK);
+			StringBuilder result = new StringBuilder();
+			
+			
+			
+			result.append("User: ").append(feed.getUser().getEmail()).append("\n").append("Feed_Id: ")
+					.append(feed.getFid()).append("\nFeed 작성완료");
+			return new ResponseEntity<Object>(result, HttpStatus.OK);
 		} catch (Exception e) {
 			throw e;
 		}
@@ -104,13 +108,50 @@ public class FeedController {
 	}
 
 	@ApiImplicitParams({
-		@ApiImplicitParam(name = "TOKEN", value = "로그인 성공 후 access_token", required = true, dataType = "String", paramType = "header") })
+			@ApiImplicitParam(name = "TOKEN", value = "로그인 성공 후 access_token", required = true, dataType = "String", paramType = "header") })
+	@ApiOperation(value = "Feed 단일 가져오기")
+	@GetMapping("/findById")
+	public ResponseEntity<Object> findById(@RequestParam Long fid) {
+		try {
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			String uid = authentication.getName();
+			FeedOutputDto feed = feedService.findById(fid);
+
+			StringBuilder result = new StringBuilder();
+			result.append("Request User: ").append(uid).append("\nFeed UID: ").append(feed.getUser().getUid());
+
+			if (uid.equals(feed.getUser().getUid().toString())) {
+				return new ResponseEntity<Object>(feed, HttpStatus.OK);
+			} else {
+				result.append("\n검색 실패: 검색을 요청한 유저와 게시글 작성자와 다릅니다.");
+				return new ResponseEntity<Object>(result, HttpStatus.ACCEPTED);
+			}
+		} catch (Exception e) {
+			throw e;
+		}
+	}
+
+	@ApiImplicitParams({
+			@ApiImplicitParam(name = "TOKEN", value = "로그인 성공 후 access_token", required = true, dataType = "String", paramType = "header") })
 	@ApiOperation(value = "Feed 수정")
 	@PutMapping("/update")
 	public ResponseEntity<Object> update(@RequestParam Long fid, @RequestBody FeedInputDto feedInput) {
 		try {
-			Integer result = feedService.update(fid, feedInput);
-			return new ResponseEntity<Object>(result + "개 수정되었습니다.", HttpStatus.OK);
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			String uid = authentication.getName();
+			FeedOutputDto feed = feedService.findById(fid);
+
+			StringBuilder result = new StringBuilder();
+			result.append("Request User: ").append(uid).append("\nFeed UID: ").append(feed.getUser().getUid());
+
+			if (uid.equals(feed.getUser().getUid().toString())) {
+				feedService.update(fid, feedInput);
+				result.append("\n수정되었습니다.");
+				return new ResponseEntity<Object>(result, HttpStatus.OK);
+			} else {
+				result.append("\n수정 실패: 수정을 요청한 유저와 수정할 게시글 작성자와 다릅니다.");
+				return new ResponseEntity<Object>(result, HttpStatus.ACCEPTED);
+			}
 		} catch (Exception e) {
 			throw e;
 		}
@@ -126,15 +167,29 @@ public class FeedController {
 			throw e;
 		}
 	}
-	
+
 	@ApiImplicitParams({
-		@ApiImplicitParam(name = "TOKEN", value = "로그인 성공 후 access_token", required = true, dataType = "String", paramType = "header") })
+			@ApiImplicitParam(name = "TOKEN", value = "로그인 성공 후 access_token", required = true, dataType = "String", paramType = "header") })
 	@ApiOperation(value = "Feed 제거")
 	@DeleteMapping("/delete")
 	public ResponseEntity<Object> delete(@RequestParam Long fid) {
 		try {
-			feedService.delete(fid);
-			return new ResponseEntity<Object>("Feed가 제거되었습니다.", HttpStatus.OK);
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			String uid = authentication.getName();
+			FeedOutputDto feed = feedService.findById(fid);
+
+			StringBuilder result = new StringBuilder();
+			result.append("Request User: ").append(uid).append("\nFeed UID: ").append(feed.getUser().getUid());
+
+			if (uid.equals(feed.getUser().getUid().toString())) {
+				feedService.delete(fid);
+				result.append("\n삭제되었습니다.");
+				return new ResponseEntity<Object>(result, HttpStatus.OK);
+			} else {
+				result.append("\n삭제 실패: 삭제을 요청한 유저와 삭제할 게시글 작성자와 다릅니다.");
+				return new ResponseEntity<Object>(result, HttpStatus.ACCEPTED);
+			}
+
 		} catch (Exception e) {
 			throw e;
 		}
